refactor(projects): extract task dependency graph helpers
